+++ framework/CHANGELOG.md
@@ -19,6 +19,7 @@ Yii Framework 2 Change Log
 - Bug #17793: Fix inconsistent handling of null `data` attribute values in `yii\helpers\BaseHtml::renderTagAttributes()` (brandonkelly)
 - Bug #17300: Fix class-level Event handling with wildcards (Toma91)
 - Bug #17635: Fix varbinary data handling for MSSQL (toatall)
+- Enh #17817: Added `WhereConditionToDataFilter` to convert $query->where to `DataFilter` format (alhimik1986)
 
 
 2.0.31 December 18, 2019